Separate out generation routes, move routes to their own directory
